Moved config slice selectors definition to selectors.ts file
