+++ client/src/state/slices/configSlice/selectors.ts
@@ -1,0 +1,5 @@
+import { RootState } from "../../store";
+
+const selectTheme = (state: RootState) => state.config.theme;
+
+export { selectTheme };
